requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0 